Scene time in double precision
A wallpaper runs for weeks; float seconds lose sub-frame precision
within a day or two (8ms ulp at ~28h), which degrades video frame
selection and time-driven animation. FrameContext::seconds and the
delta computation are double end to end; f32 only at the value-port
boundary (@time.seconds), a documented limit of the f32 value graph.

+++ src/core/Nodes.cpp
@@ -142,9 +142,9 @@ TimeNode::TimeNode()
 void TimeNode::evaluate(FrameContext& ctx)
 {
     Value seconds{};
-    seconds.v[0] = ctx.seconds;
+    seconds.v[0] = (float)ctx.seconds;
     Value delta{};
-    delta.v[0] = firstEvaluate ? 0.0f : ctx.seconds - mLast;
+    delta.v[0] = firstEvaluate ? 0.0f : (float)(ctx.seconds - mLast);
     mLast = ctx.seconds;
     writeOutput(outputs[0], seconds);
     writeOutput(outputs[1], delta);
